requirements.txt
python-telegram-bot[job-queue]>=21.0
anthropic>=0.34.0
httpx>=0.27.0
pytest>=8.0.0
python-dotenv>=1.0.1
